requirements.txt
pydantic
pyserial
pyserial_asyncio
pyyaml
wxpython
wxasync
ObjectListView
logzero
